requirements.txt
tensorflow
numpy
opencv-python
scikit-learn
streamlit
Pillow
faiss-cpu